+++ dask_geomodeling/tests/test_geometry.py
@@ -151,6 +151,7 @@ class TestGeometryFileSource(unittest.TestCase):
         bbox3857 = extent.transformed(get_sr("EPSG:3857")).bbox
         result = self.source.get_data(geometry=box(*bbox3857), projection="EPSG:3857")
         self.assertEqual("EPSG:3857", result["projection"])
+        self.assertEqual("epsg:3857", result["features"].crs["init"])
         self.assertEqual(10, len(result["features"]))
 
     def test_limit(self):